Fix handling of trailing space in origin-field
We use the IOI trk-101. This device returns an origin-field with a
trailing space. This triggered the special case for some onvif2 cameras.
This resulted in a double unicast-address, causing a shift in the
fields. Eventually resulting in an exception when nettype was
interpret as sess-version.

This commit fixes this issue and adds a testcase that includes a
trailing whitespace.