Allow setting response body in OnDescribe (#758)
* Allow setting response body in OnDescribe

Since commit 8c6495c33b3194f1bf34f3f8a87f51907e567915, the server can no
longer directly specify the response body to DESCRIBE requests. This is
a breaking change and is a limitation for users who manually handle their
SDPs. This commit allows setting the body explicitly and takes it into
account if the `ServerStream` part of the `OnDecsribe` return value is
nil. If neither the stream, nor the body are set, the application panics.

* Add test for non-nil body for Describe